backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
azure-ai-formrecognizer==3.3.2
azure-ai-textanalytics==5.3.0
pydantic[email]==2.5.2
